+++ src/me/anno/gpu/pipeline/InstancedStackStatic.kt
@@ -71,7 +71,7 @@ class InstancedStackStatic(capacity: Int = 512) : KeyPairMap<Mesh, Material, Ins
                         GFX.shaderColor(shader, "tint", -1)
                         shader.v1b("hasAnimation", false)
                         shader.v1i("hasVertexColors", mesh.hasVertexColors)
-                        shader.v2i("randomIdData", mesh.numPrimitives, 0)
+                        shader.v2i("randomIdData", mesh.numPrimitives.toInt(), 0)
                         GFX.check()
 
                         shader.v4f("clickId", stack.clickIds[i])
